% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{longest_common_substring}
\alias{longest_common_substring}
\title{Find the longest common substring of two strings.}
\usage{
longest_common_substring(str1, str2)
}
\arguments{
\item{str1}{A character vector}

\item{str2}{A character vector}
}
\description{
Find the longest common substring of two strings.
}
